- Swap SWL with Drawndown - Removed redundant activity submission code (from django forms) - Updated api tests - Get rid of productiondata - Updated migration - Simplified gwells.forms - Update export to work with new production data location - Update replication to work with new production data location - Added recommended pump depth & rate (for edit view only) - Update fixtures - Added default rate unit of measure
