+++ app/backend/gwells/views/__init__.py
@@ -13,48 +13,10 @@
 """
 
 from ..forms import *
-from gwells.views.activity_submission import ActivitySubmissionDetailView, ActivitySubmissionListView,\
-    ActivitySubmissionWizardView
 from gwells.views.health_view import HealthView
 from gwells.views.registry_view import RegistryView
 from gwells.views.search_view import SearchView
 from gwells.views.survey_list_view import SurveyListView
 
-FORMS = [('type_and_class', ActivitySubmissionTypeAndClassForm),
-         ('owner', WellOwnerForm),
-         ('location', ActivitySubmissionLocationForm),
-         ('gps', ActivitySubmissionGpsForm),
-         ('lithology', ActivitySubmissionLithologyFormSet),
-         ('casing', ActivitySubmissionCasingFormSet),
-         ('surface_seal', ActivitySubmissionSurfaceSealForm),
-         ('liner_perforation', ActivitySubmissionLinerPerforationFormSet),
-         ('screen_intake', ActivitySubmissionScreenIntakeForm),
-         ('screen', ActivitySubmissionScreenFormSet),
-         ('filter_pack', ActivitySubmissionFilterPackForm),
-         ('development', ActivitySubmissionDevelopmentForm),
-         ('production_data', ProductionDataFormSet),
-         ('water_quality', ActivitySubmissionWaterQualityForm),
-         ('well_completion', WellCompletionForm),
-         ('comments', ActivitySubmissionCommentForm),
-         ]
 
-TEMPLATES = {'type_and_class': 'gwells/activity_submission_form.html',
-             'owner': 'gwells/activity_submission_form.html',
-             'location': 'gwells/activity_submission_form.html',
-             'gps': 'gwells/activity_submission_form.html',
-             'lithology': 'gwells/activity_submission_lithology_form.html',
-             'casing': 'gwells/activity_submission_casing_form.html',
-             'surface_seal': 'gwells/activity_submission_form.html',
-             'liner_perforation': 'gwells/activity_submission_liner_perforation_form.html',
-             'screen_intake': 'gwells/activity_submission_form.html',
-             'screen': 'gwells/activity_submission_screen_form.html',
-             'filter_pack': 'gwells/activity_submission_form.html',
-             'development': 'gwells/activity_submission_form.html',
-             'production_data': 'gwells/activity_submission_form.html',
-             'water_quality': 'gwells/activity_submission_form.html',
-             'well_completion': 'gwells/activity_submission_form.html',
-             'comments': 'gwells/activity_submission_form.html',
-             }
-
-__all__ = ['ActivitySubmissionDetailView', 'ActivitySubmissionListView', 'ActivitySubmissionWizardView',
-           'FORMS', 'HealthView', 'RegistryView', 'SearchView', 'TEMPLATES']
+__all__ = ['HealthView', 'RegistryView', 'SearchView']
