+++ app/backend/wells/models.py
@@ -500,6 +500,8 @@ class CoordinateAcquisitionCode(AuditModel):
         return self.description
 
 
+# TODO: Consider having Well and Submission extend off a common base class, given that
+#   they mostly have the exact same fields!
 class Well(AuditModel):
     """
     Well information.
@@ -808,6 +810,38 @@ class Well(AuditModel):
         max_digits=5, decimal_places=2, blank=True, null=True, verbose_name='Analytic Solution Type')
     boundary_effect = models.DecimalField(
         max_digits=5, decimal_places=2, blank=True, null=True, verbose_name='Boundary Effect')
+
+    # Production data related data
+    yield_estimation_method = models.ForeignKey(
+        YieldEstimationMethodCode, db_column='yield_estimation_method_code',
+        on_delete=models.CASCADE, blank=True, null=True,
+        verbose_name='Estimation Method')
+    yield_estimation_rate = models.DecimalField(
+        max_digits=7, decimal_places=2, verbose_name='Estimation Rate',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.00'))])
+    yield_estimation_duration = models.DecimalField(
+        max_digits=9, decimal_places=2, verbose_name='Estimation Duration',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.01'))])
+    static_level_before_test = models.DecimalField(
+        max_digits=7, decimal_places=2, verbose_name='SWL Before Test',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.0'))])
+    drawdown = models.DecimalField(
+        max_digits=7, decimal_places=2, blank=True, null=True,
+        validators=[MinValueValidator(Decimal('0.00'))])
+    hydro_fracturing_performed = models.BooleanField(
+        default=False, verbose_name='Hydro-fracturing Performed?',
+        choices=((False, 'No'), (True, 'Yes')))
+    hydro_fracturing_yield_increase = models.DecimalField(
+        max_digits=7, decimal_places=2,
+        verbose_name='Well Yield Increase Due to Hydro-fracturing',
+        blank=True, null=True,
+        validators=[MinValueValidator(Decimal('0.00'))])
+    recommended_pump_depth = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
+                                                 verbose_name='Recommended pump depth',
+                                                 validators=[MinValueValidator(Decimal('0.00'))])
+    recommended_pump_rate = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
+                                                verbose_name='Recommended pump rate',
+                                                validators=[MinValueValidator(Decimal('0.00'))])
 
     class Meta:
         db_table = 'well'
@@ -1233,6 +1267,38 @@ class ActivitySubmission(AuditModel):
     boundary_effect = models.DecimalField(
         max_digits=5, decimal_places=2, blank=True, null=True, verbose_name='Boundary Effect')
 
+    # Production data related data
+    yield_estimation_method = models.ForeignKey(
+        YieldEstimationMethodCode, db_column='yield_estimation_method_code',
+        on_delete=models.CASCADE, blank=True, null=True,
+        verbose_name='Estimation Method')
+    yield_estimation_rate = models.DecimalField(
+        max_digits=7, decimal_places=2, verbose_name='Estimation Rate',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.00'))])
+    yield_estimation_duration = models.DecimalField(
+        max_digits=9, decimal_places=2, verbose_name='Estimation Duration',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.01'))])
+    static_level_before_test = models.DecimalField(
+        max_digits=7, decimal_places=2, verbose_name='SWL Before Test',
+        blank=True, null=True, validators=[MinValueValidator(Decimal('0.0'))])
+    drawdown = models.DecimalField(
+        max_digits=7, decimal_places=2, blank=True, null=True,
+        validators=[MinValueValidator(Decimal('0.00'))])
+    hydro_fracturing_performed = models.BooleanField(
+        default=False, verbose_name='Hydro-fracturing Performed?',
+        choices=((False, 'No'), (True, 'Yes')))
+    hydro_fracturing_yield_increase = models.DecimalField(
+        max_digits=7, decimal_places=2,
+        verbose_name='Well Yield Increase Due to Hydro-fracturing',
+        blank=True, null=True,
+        validators=[MinValueValidator(Decimal('0.00'))])
+    recommended_pump_depth = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
+                                                 verbose_name='Recommended pump depth',
+                                                 validators=[MinValueValidator(Decimal('0.00'))])
+    recommended_pump_rate = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
+                                                verbose_name='Recommended pump rate',
+                                                validators=[MinValueValidator(Decimal('0.00'))])
+
     class Meta:
         db_table = 'activity_submission'
 
@@ -1319,67 +1385,6 @@ class LithologyDescription(AuditModel):
                                                          self.lithology_to)
         else:
             return 'well {} {} {}'.format(self.well, self.lithology_from, self.lithology_to)
-
-
-class ProductionData(AuditModel):
-    """
-    Water production of a well measured by a driller
-    """
-    production_data_guid = models.UUIDField(
-        primary_key=True, default=uuid.uuid4, editable=False)
-    activity_submission = models.ForeignKey(
-        ActivitySubmission, db_column='filing_number',
-        on_delete=models.CASCADE, blank=True, null=True)
-    well = models.ForeignKey(
-        Well, db_column='well_tag_number', on_delete=models.CASCADE,
-        blank=True, null=True)
-    yield_estimation_method = models.ForeignKey(
-        YieldEstimationMethodCode, db_column='yield_estimation_method_code',
-        on_delete=models.CASCADE, blank=True, null=True,
-        verbose_name='Estimation Method')
-    yield_estimation_rate = models.DecimalField(
-        max_digits=7, decimal_places=2, verbose_name='Estimation Rate',
-        blank=True, null=True, validators=[MinValueValidator(Decimal('0.00'))])
-    yield_estimation_duration = models.DecimalField(
-        max_digits=9, decimal_places=2, verbose_name='Estimation Duration',
-        blank=True, null=True, validators=[MinValueValidator(Decimal('0.01'))])
-    well_yield_unit = models.ForeignKey(
-        WellYieldUnitCode, db_column='well_yield_unit_code', blank=True,
-        null=True, on_delete=models.PROTECT)
-    static_level = models.DecimalField(
-        max_digits=7, decimal_places=2, verbose_name='SWL Before Test',
-        blank=True, null=True, validators=[MinValueValidator(Decimal('0.0'))])
-    drawdown = models.DecimalField(
-        max_digits=7, decimal_places=2, blank=True, null=True,
-        validators=[MinValueValidator(Decimal('0.00'))])
-    hydro_fracturing_performed = models.BooleanField(
-        default=False, verbose_name='Hydro-fracturing Performed?',
-        choices=((False, 'No'), (True, 'Yes')))
-    hydro_fracturing_yield_increase = models.DecimalField(
-        max_digits=7, decimal_places=2,
-        verbose_name='Well Yield Increase Due to Hydro-fracturing',
-        blank=True, null=True,
-        validators=[MinValueValidator(Decimal('0.00'))])
-
-    recommended_pump_depth = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
-                                                verbose_name='Recommended pump depth',
-                                                validators=[MinValueValidator(Decimal('0.00'))])
-    recommended_pump_rate = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True,
-                                                verbose_name='Recommended pump rate',
-                                                validators=[MinValueValidator(Decimal('0.00'))])
-
-    class Meta:
-        db_table = 'production_data'
-
-    def __str__(self):
-        if self.activity_submission:
-            return 'activity_submission {} {} {}'.format(
-                self.activity_submission, self.yield_estimation_method,
-                self.yield_estimation_rate)
-        else:
-            return 'well {} {} {}'.format(
-                self.well, self.yield_estimation_method,
-                self.yield_estimation_rate)
 
 
 class LinerPerforation(AuditModel):
