+++ app/frontend/src/submissions/components/SubmissionForm/ActivitySubmissionForm.vue
@@ -308,8 +308,16 @@ Licensed under the Apache License, Version 2.0 (the "License");
       <yield class="my-5"
         v-if="showSection('wellYield')"
         id="wellYield"
-        :productionData.sync="form.productiondata_set"
         :errors="errors"
+        :yieldEstimationMethod.sync="form.yield_estimation_method"
+        :yieldEstimationRate.sync="form.yield_estimation_rate"
+        :yieldEstimationDuration.sync="form.yield_estimation_duration"
+        :staticLevel.sync="form.static_level_before_test"
+        :drawdown.sync="form.drawdown"
+        :hydroFracturingPerformed.sync="form.hydro_fracturing_performed"
+        :hydroFracturingYieldIncrease.sync="form.hydro_fracturing_yield_increase"
+        :recommendedPumpDepth.sync="form.recommended_pump_depth"
+        :recommendedPumpRate.sync="form.recommended_pump_rate"
         :isStaffEdit="isStaffEdit"
         :saveDisabled="editSaveDisabled"
         v-on:save="$emit('submit_edit')"
